pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "feedparser>=6.0.11",
    "openai>=1.69.0",
    "pandas>=2.2.3",
    "python-dateutil>=2.9.0.post0",
    "schedule>=1.2.2",
    "streamlit>=1.44.0",
]
